backend/requirements.txt
# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Graph processing
networkx==3.2.1

# Data models
pydantic==2.5.0

# Scheduling
APScheduler==3.10.4

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2

# Report generation
reportlab==4.0.7

# Google services
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# AI/LLM services
openai==1.3.8
google-generativeai==0.3.2
anthropic==0.7.8

# Communication services
twilio==8.10.3

# Image processing
Pillow==10.1.0